refactor(desktop): extract usePanZoom hook from ImageView
