+++ apps/desktop/src/renderer/routes/_authenticated/_dashboard/v2-workspace/$workspaceId/hooks/usePaneRegistry/components/FilePane/registry/views/ImageView/hooks/usePanZoom/index.ts
@@ -1,0 +1,1 @@
+export { usePanZoom } from "./usePanZoom";
